+++ pakyow-support/CHANGELOG.md
@@ -1,5 +1,8 @@
 # v1.1.0 (unreleased)
 
+  * `add` **Allow object name to be passed explicitly when isolating an object.**
+    - [Pull Request #369][pr-369]
+
   * `chg` **Pass isolable namespace as a keyword argument.**
     - [Pull Request #368][pr-368]
 
@@ -92,6 +95,7 @@
     *Related links:*
     - [Pull Request #364][pr-364]
 
+[pr-369]: https://github.com/pakyow/pakyow/pull/369
 [pr-368]: https://github.com/pakyow/pakyow/pull/368
 [pr-367]: https://github.com/pakyow/pakyow/pull/367
 [pr-366]: https://github.com/pakyow/pakyow/pull/366